js files moved into src and all string concats converted to template literals.
