+++ package.json
@@ -6,6 +6,7 @@
   "author": "Kent",
   "dependencies": {
     "discord.js": "git+https://github.com/hydrabolt/discord.js.git",
+    "lebab": "^2.7.7",
     "lodash.get": "^4.4.2",
     "util-inspect": "^0.1.8",
     "winston": "^2.3.1"
